### 2019-12-06 #1936 @whedon
[PRE REVIEW]: EspressoDB: A scientific database for managing high-performance computing workflows
**Submitting author:** @ckoerber (<a href="http://orcid.org/0000-0002-9271-8022">Christopher Körber</a>)
**Repository:** <a href="https://github.com/callat-qcd/espressodb" target ="_blank">https://github.com/callat-qcd/espressodb</a>
**Version:** v1.0.0
