+++ doc-src/arXiv/arXiv.tex
@@ -1,7 +1,7 @@
 % -------------------- Acknowledgement -------------------------
 %
 % This latex file is copied from that prepared by Matthew Pitkin
-% for his JOSS article: 
+% for his JOSS article:
 %     psrqpy: a python interface for querying the ATNF pulsar catalogue
 %     https://joss.theoj.org/papers/10.21105/joss.00538
 %     https://arxiv.org/abs/1806.07809
@@ -190,7 +190,7 @@
         \author[1,2,3]{Chia Cheng Chang}
         \author[2,3]{Christopher K{\"o}rber}
         \author[3,2]{Andr{\'e}~Walker-Loud}
-    
+
       \affil[1]{iTHEMS RIKEN, Wako, Saitama 351-0198, Japan}
       \affil[2]{Department of Physics, University of California, Berkeley, California 94720, USA}
       \affil[3]{Nuclear Science Division, Lawrence Berkeley National Laboratory, Berkeley, California 94720}
