### 2019-12-07 c7175a8
prep use-case.md for supplemental section
