### 2019-12-07 1505c9b
Added new arxiv docs
