+++ doc-src/arXiv/use-case.tex
@@ -1,0 +1,209 @@
+\hypertarget{django-overview}{%
+\section{Django overview}\label{django-overview}}
+
+\texttt{EspressoDB} utilizes Python's \texttt{Django} Object-Relational
+Mapping (ORM) framework. Tables correspond to Python classes, rows
+correspond to instances and columns correspond to attributes of the
+instances. Thus it is possible to filter for objects by their attributes
+or generate summary tables (\texttt{pandas.DataFrame}) within one line
+of code. Furthermore, using an ORM allows one to have the same interface
+independent of the backend. It is possible to store data in a file based
+\texttt{SQLite} solution, or use more scalable options like
+\texttt{MySQL} or \texttt{Postgresql}.
+
+\texttt{Django} is part of many open-source projects and thus comes with
+extensive documentation. Additionally, \texttt{Django} is scalable,
+comes with reliable tests and vast community support which manifests in
+the fact that it is commonly used in large scale projects (BitBucket,
+Instagram, Mozilla, NASA and many more). One guiding principle of
+\texttt{EspressoDB} is to not ``re-invent the wheel'' but instead
+leverage the support coming from \texttt{Django}. As a result, one can
+easily incorporate many of \texttt{Django}'s extensions and find
+solutions to technical questions online.
+
+\hypertarget{lattice-qcd-use-case}{%
+\section{Lattice QCD use case}\label{lattice-qcd-use-case}}
+
+LQCD is an inherently a stochastic method of simulating quantum
+chromodynamics (QCD) the fundamental theory of nuclear strong
+interactions, which is responsible for confining quarks into protons and
+neutrons and ultimately, for binding these nucleons into the atomic
+nuclei we observe in nature. The application of LQCD to forefront
+research applications in nuclear physics is necessary to build a
+quantitative connection between our understanding of nuclei and QCD.
+This is important as nuclei serve as laboratories and/or detectors for
+many experiments aiming to test the limits of the Standard Model of
+particle physics in our quest to understand questions such as: Why is
+the universe composed of matter and not anti-matter? Does dark matter
+interact with regular matter other than gravitationally? What is the
+nature of the neutrino and is it related to the observed excess of
+matter over anti-matter? See Ref. \cite{Drischler:2019xuo} for a recent
+review of the interface of LQCD with our understanding of nuclear
+physics.
+
+The application of LQCD to nuclear physics is an exascale challenge. One
+of the main reasons these calculations are so expensive is that when
+LQCD is applied to systems of one or more nucleons, an exponentially bad
+signal-to-noise problem must be overcome. While the optimal strategy for
+overcoming this challenge is not yet known, one thing common to all
+methods is the need for an exponentially large amount of statistics. As
+such, these LQCD computations require the completion of millions of
+independent sub-calculations (or tasks), with chained dependencies, in
+order to complete a single calculation. These chained tasks write large
+amounts of temporary files to the scratch file system which are used as
+input for subsequent files, often with multiple input files required for
+the later computations. Several such calculations must be performed in
+order to extrapolate the results to the physical point, defined as the
+limit of zero discretization (the continuum limit), the infinite volume
+limit and the limit of physical quark masses which are \emph{a priori}
+unknown and so must be determined through this
+extrapolation/interpolation procedure. These requirements lead to a very
+complex set of computations that must be carried out with great care and
+a significant amount of human time and effort to ensure the computing
+cycles are used as efficiently as possible.
+
+Compounding these challenges, the new computer Summit, is
+\emph{disruptively fast} compared to previous generations of leadership
+class computers. Full machine-to-machine, Summit is approximately 15
+times faster than Titan when applied to LQCD applications such as those
[…]
