+++ Makefile
@@ -2,7 +2,7 @@ paper | paper.pdf: paper.md paper.bib
 	pandoc -t latex -o paper.pdf --bibliography paper.bib paper.md
 
 tex: paper.md paper.bib
-	pandoc -t latex -o doc-src/arXiv/paper_body.tex --bibliography paper.bib paper.md
+	pandoc -t latex -o doc-src/arXiv/paper_body.tex paper.md
 	sed -i.bak 's:doc-src/_static/lattedb-example.png:lattedb-example.png:' doc-src/arXiv/paper_body.tex
 	rm doc-src/arXiv/paper_body.tex.bak
 	cp doc-src/_static/lattedb-example.png doc-src/arXiv/
