### 2019-12-07 cb21b20
Added new arxiv docs
