+++ doc-src/arXiv/arXiv.tex
@@ -229,6 +229,7 @@
 
 
 \input{paper_body}
+\input{use-case}
 
 
 \end{document}
