### 2019-12-07 244bba2
working paper.md to doc-src/arXiv/paper_body.tex
