+++ doc-src/arXiv/arXiv.tex
@@ -5,6 +5,8 @@
 %     psrqpy: a python interface for querying the ATNF pulsar catalogue
 %     https://joss.theoj.org/papers/10.21105/joss.00538
 %     https://arxiv.org/abs/1806.07809
+%
+% With modifications to the bibtex, hypersetup
 % --------------------------------------------------------------
 \documentclass[10pt,a4paper,onecolumn]{article}
 \usepackage{marginnote}
@@ -15,21 +17,25 @@
 \usepackage{calc}
 \usepackage{tikz}
 \usepackage{hyperref}
-\hypersetup{colorlinks,breaklinks,
-            urlcolor=[rgb]{0.0, 0.5, 1.0},
-            linkcolor=[rgb]{0.0, 0.5, 1.0}}
+\hypersetup{
+    colorlinks=true,     % false: boxed links; true: colored links
+    linkcolor=blue,      % color of internal links
+    citecolor=blue,      % color of links to bibliography
+    filecolor=blue,      % color of file links
+    urlcolor=blue        % color of external links
+}
 \usepackage{caption}
 \usepackage{tcolorbox}
 \usepackage{amssymb,amsmath}
 \usepackage{ifxetex,ifluatex}
 \usepackage{seqsplit}
-\usepackage{fixltx2e} % provides \textsubscript
-\usepackage[
-  backend=biber,
+%\usepackage{fixltx2e} % provides \textsubscript
+%\usepackage[
+%  backend=bibtex,
 %  style=alphabetic,
 %  citestyle=numeric
-]{biblatex}
-\bibliography{paper.bib}
+%]{biblatex}
+%\bibliography{paper.bib}
 
 
 % --- Page layout -------------------------------------------------------------
@@ -40,7 +46,7 @@
 % \renewcommand\familydefault{\sfdefault}
 
 % --- Style -------------------------------------------------------------------
-\renewcommand{\bibfont}{\small \sffamily}
+%\renewcommand{\bibfont}{\small \sffamily}
 \renewcommand{\captionfont}{\small\sffamily}
 \renewcommand{\captionlabelfont}{\bfseries}
 
@@ -67,9 +73,8 @@
 \fancyhead[L]{\hspace{-0.75cm}\includegraphics[width=5.5cm]{joss-logo.png}}
 \fancyhead[C]{}
 \fancyhead[R]{}
-\renewcommand{\footrulewidth}{0.25pt}
+%\renewcommand{\footrulewidth}{0.25pt}
 
-\fancyfoot[L]{\footnotesize{\sffamily Pitkin, (2018). psrqpy: a python interface for querying the ATNF pulsar catalogue. \textit{Journal of Open Source Software}, 3(22), 538. \href{https://doi.org/10.21105/joss.00538}{https://doi.org/10.21105/joss.00538}}}
 
 
 \fancyfoot[R]{\sffamily \thepage}
@@ -150,7 +155,7 @@
 \UseMicrotypeSet[protrusion]{basicmath} % disable protrusion for tt fonts
 }{}
 
-\usepackage{hyperref}
+%\usepackage{hyperref}
 \hypersetup{unicode=true,
             pdftitle={psrqpy: a python interface for querying the ATNF pulsar catalogue},
             pdfborder={0 0 0},
@@ -227,9 +232,18 @@
   Authors of papers retain copyright and release the work under a Creative Commons Attribution 4.0 International License (\href{http://creativecommons.org/licenses/by/4.0/}{\color{linky}{CC-BY}}).
 }
 
-
+%---------------------------------------------------------
+% Main body
 \input{paper_body}
 \input{use-case}
+
+%---------------------------------------------------------
+% bib files loaded at top
+
+\bibliography{paper}
+\bibliographystyle{ieeetr}
+
+
 
 
 \end{document}
