### 2019-12-07 abc506a
working paper.md to doc-src/arXiv/paper_body.tex
