### 2019-12-07 2254f1e
Removed paper_body for for_real
