+++ doc-src/arXiv/paper_body.tex
@@ -1,160 +1,0 @@
-\hypertarget{summary}{%
-\section{Summary}\label{summary}}
-
-Leadership computing facilities around the world support cutting-edge
-scientific research across a broad spectrum of disciplines including
-understanding climate change \cite{Kurth\_2018}, combating opioid
-addiction \cite{Joubert:2018:AOE:3291656.3291732}, or simulating the
-decay of a neutron \cite{Berkowitz:2018gqe}. While the increase in
-computational power has allowed scientists to better evaluate the
-underlying model, the size of these computational projects have grown to
-a point where a framework is desired to facilitate managing the
-workflow. A typical scientific computing workflow includes:
-
-\begin{enumerate}
-\def\labelenumi{\arabic{enumi}.}
-\tightlist
-\item
-  Defining all input parameters for every step of the computation;
-\item
-  Defining dependencies of computational tasks;
-\item
-  Storing some of the output data;
-\item
-  Post-processing these data files;
-\item
-  Performing data analysis on output.
-\end{enumerate}
-
-\href{https://github.com/callat-qcd/espressodb/}{\texttt{EspressoDB}} is
-a programmatic object-relational data management framework implemented
-in Python and based on the
-\href{https://www.djangoproject.com}{\texttt{Django} web framework}.
-\texttt{EspressoDB} was developed to streamline data management
-workflows, centralize and guarantee data integrity, while providing
-domain flexibility and ease of use.
-
-The framework provided by \texttt{EspressoDB} aims to support the ever
-increasing complexity of workflows of scientific computing at leadership
-computing facilities (LCFs), with the goal of reducing the amount of
-human time required to manage the jobs, thus giving scientists more time
-to focus on science.
-
-\hypertarget{features}{%
-\section{Features}\label{features}}
-
-Data integrity is important to scientific projects and becomes more
-challenging the larger the project. In general, a \texttt{SQL} framework
-type-checks data before writing to the database and controls
-dependencies and relations between different tables to ensure internal
-consistency. \texttt{EspressoDB} allows additional user-defined
-constraints not supported by \texttt{SQL} (\emph{e.g.} unique
-constraints using information across related tables). Once the user has
-specified a set of conditions that entries have to fulfill for each
-table, \texttt{EspressoDB} runs these cross checks for new data before
-inserting them in the database.
-
-\texttt{EspressoDB} also supports collaborative and open-data oriented
-projects by leveraging and extending \texttt{Django}'s web hosting
-component. In addition to providing a centralized data platform, it is
-possible to spawn customized web pages which can be hosted locally or on
-the world wide web\footnote{Depending on the configuration, it is
-  possible to provide selected access for multiple users on different
-  levels.}. \texttt{EspressoDB} simplifies creating projects by
-providing default \texttt{Django} configurations that set up for
-example, connections to the database and webpages to view associated
-tables. For example, with the default setting, \texttt{EspressoDB}
-spawns:
-
-\begin{itemize}
-\tightlist
-\item
-  Documentation views of implemented tables;
-\item
-  A project wide notification system;
-\item
-  Project specific Python interface guidelines which help writing
-  scripts to populate the database;
-\item
-  Admin pages for interacting with data in a GUI.
-\end{itemize}
-
-Further views can be implemented to interact with data and use existing
-Python libraries for summarizing and visualizing information. This
-allows users to create visual progress updates on the fly and to
-integrate the database information to the data-processing workflow,
[…]
