### 2019-12-07 a604263
moving use case to appendix - other small changes
