+++ doc-src/arXiv/use-case.tex
@@ -176,11 +176,12 @@ to manage the computations. We are actively constructing routines to
 also store the final data files to tape, the status of which is stored
 in a related \texttt{LatteDB} table. Thus, the user can query the
 database instead of the file system for the existence of data files,
-significantly reducing the load on the file system as well. Example use
-scripts for interacting with \texttt{LatteDB} can be found in our
-management repository for these INCITE projects
-\url{https://github.com/callat-qcd/nucleon_elastic_FF} in the
-\texttt{scripts} folder and the \texttt{notebooks} folder in
+significantly reducing the load on the file system as well.
+Examples of interacting with \texttt{LatteDB} can be found in
+our management repository for these INCITE projects
+\url{https://github.com/callat-qcd/nucleon_elastic_FF}.
+The scripts for interacting with \texttt{LatteDB} are in the \texttt{scripts} folder and containt \texttt{lattedb} in the name.
+Other examples can be found in the \texttt{notebooks} folder in
 \texttt{LatteDB}. These scripts will be updated regularly to encompass
 more and more utilities from \texttt{EspressoDB} providing a complete
 working example.
