+++ doc-src/arXiv/arXiv.tex
@@ -209,7 +209,7 @@
   %\hrule
   \sffamily\small
 
-  {\bfseries DOI:} \href{XXX}{\color{linky}{XXX}}
+  {\bfseries DOI:} \href{XXX}{\color{linky}{TBD}}
 
   \vspace{2mm}
 
@@ -235,7 +235,6 @@
 %---------------------------------------------------------
 % Main body
 \input{paper_body}
-\input{use-case}
 
 %---------------------------------------------------------
 % bib files loaded at top
@@ -243,6 +242,8 @@
 \bibliography{paper}
 \bibliographystyle{ieeetr}
 
+\appendix
+\input{use-case}
 
 
 
