### 2019-12-07 1d23335
moving use case to appendix - other small changes
