### 2019-12-07 1b03e91
added Supplemental to start of use case, demoted sections to subsections in use-case; fixed comments in arXiv
