+++ doc-src/arXiv/arXiv.tex
@@ -237,7 +237,7 @@
 \input{paper_body}
 
 %---------------------------------------------------------
-% bib files loaded at top
+% bib file
 
 \bibliography{paper}
 \bibliographystyle{ieeetr}
