+++ doc-src/arXiv/use-case.tex
@@ -1,5 +1,6 @@
-\hypertarget{django-overview}{%
-\section{Django overview}\label{django-overview}}
+\section{Supplemental Information}
+
+\subsection{Django overview\label{django-overview}}
 
 \texttt{EspressoDB} utilizes Python's \texttt{Django} Object-Relational
 Mapping (ORM) framework. Tables correspond to Python classes, rows
@@ -21,8 +22,7 @@ leverage the support coming from \texttt{Django}. As a result, one can
 easily incorporate many of \texttt{Django}'s extensions and find
 solutions to technical questions online.
 
-\hypertarget{lattice-qcd-use-case}{%
-\section{Lattice QCD use case}\label{lattice-qcd-use-case}}
+\subsection{Lattice QCD use case \label{lattice-qcd-use-case}}
 
 LQCD is an inherently a stochastic method of simulating quantum
 chromodynamics (QCD) the fundamental theory of nuclear strong
