### 2019-12-07 983d7aa
added Supplemental to start of use case, demoted sections to subsections in use-case; fixed comments in arXiv
