+++ doc-src/arXiv/arXiv.tex
@@ -239,7 +239,7 @@
 \input{paper_body}
 
 %---------------------------------------------------------
-% bib files loaded at top
+% bib file
 
 \bibliography{paper}
 \bibliographystyle{ieeetr}
