+++ doc-src/arXiv/use-case.tex
@@ -188,7 +188,7 @@ working example.
 
 \begin{figure}
 \centering
-\includegraphics{../_static/lattedb-example.png}
+\includegraphics{lattedb-example.png}
 \caption{Example table view of file status with column specific filters
 and dynamic progress bar.}
 \end{figure}
