+++ doc-src/arXiv/arXiv.tex
@@ -157,7 +157,7 @@
 
 %\usepackage{hyperref}
 \hypersetup{unicode=true,
-            pdftitle={psrqpy: a python interface for querying the ATNF pulsar catalogue},
+            pdftitle={EspressoDB: A scientific database for managing high-performance computing workflows.},
             pdfborder={0 0 0},
             breaklinks=true}
 \urlstyle{same}  % don't use monospace font for urls
@@ -216,11 +216,13 @@
   {\bfseries Software}
   \begin{itemize}
     \setlength\itemsep{0em}
-    \item \href{XXX}{\color{linky}{Review}} \ExternalLink
-    \item \href{https://github.com/callat-qcd/espressodb}{\color{linky}{Repository}} \ExternalLink
-    \item \href{https://github.com/callat-qcd/lattedb}{\color{linky}{Repository}} \ExternalLink
-    \item \href{XXX}{\color{linky}{Archive}} \ExternalLink
+    \item \href{https://github.com/openjournals/joss-reviews/issues/1936}{\color{linky}{Review}} \ExternalLink
+    \item \href{https://github.com/callat-qcd/espressodb}{\color{linky}{EspressoDB Repository}} \ExternalLink
+    \item \href{https://github.com/callat-qcd/lattedb}{\color{linky}{LattedDB Repository}} \ExternalLink
+    \item \href{https://ithems.lbl.gov/lattedb}{\color{linky}{LattedDB Webpage}} \ExternalLink
   \end{itemize}
+
+
 
   \vspace{2mm}
 
