+++ doc-src/arXiv/arXiv.tex
@@ -193,14 +193,17 @@
 \title{EspressoDB: A scientific database for managing high-performance computing workflow}
 
         \author[1,2,3]{Chia Cheng Chang}
-        \author[2,3]{Christopher K{\"o}rber}
+        \author[2,3$\dagger$]{Christopher K{\"o}rber}
         \author[3,2]{Andr{\'e}~Walker-Loud}
 
       \affil[1]{iTHEMS RIKEN, Wako, Saitama 351-0198, Japan}
       \affil[2]{Department of Physics, University of California, Berkeley, California 94720, USA}
       \affil[3]{Nuclear Science Division, Lawrence Berkeley National Laboratory, Berkeley, California 94720}
+      \affil[$\dagger$]{\textit {\href{mailto:ckoerber@berkeley.edu}{ckoerber@berkeley.edu}}}
 
   \date{\vspace{-5ex}}
+
+
 
 \begin{document}
 \maketitle
