+++ paper.md
@@ -72,7 +72,7 @@ More details, usage instructions, and examples are documented at [espressodb.rea
 
 
 # Use case
-[``LatteDB``](https://github.com/callat-qcd/lattedb/), an application of ``EspressoDB`` that is specialized to contain table definitions for lattice quantum chromodynamics (LQCD) calculations and analysis.
+[``LatteDB``](https://github.com/callat-qcd/lattedb/) is an application of ``EspressoDB`` that is specialized to contain table definitions for lattice quantum chromodynamics (LQCD) calculations and analysis.
 ``LatteDB`` is currently being used by the [CalLat Collaboration](https://a51.lbl.gov/~callat/webhome/) in their computations on Summit at the Oak Ridge Leadership Computing Facility ([OLCF](https://www.olcf.ornl.gov)) through DOE INCITE Allocations [@incite:2019; @incite:2020].
 The website generated by ``LatteDB`` used by CalLat can be found at [https://ithems.lbl.gov/lattedb/](https://ithems.lbl.gov/lattedb/). A precursor to ``EspressoDB`` and ``LatteDB`` was used to support a series of LQCD projects [@Nicholson:2018mwc; @Chang:2018uxx].
 
