### 2019-12-07 f9f7804
updated Makefile to create tgz file to upload to arXiv
