### 2019-12-07 899764a
updated Makefile to create tgz file to upload to arXiv
