+++ doc-src/arXiv/arXiv.tex
@@ -229,7 +229,7 @@
 
   \vspace{2mm}
 
-  {\bfseries Submitted:} XX December 2019\\
+  {\bfseries Submitted:} 06 December 2019\\
   {\bfseries Published:} XX XXX 20XX
 
   \vspace{2mm}
