+++ doc-src/arXiv/arXiv.tex
@@ -199,7 +199,7 @@
       \affil[1]{iTHEMS RIKEN, Wako, Saitama 351-0198, Japan}
       \affil[2]{Department of Physics, University of California, Berkeley, California 94720, USA}
       \affil[3]{Nuclear Science Division, Lawrence Berkeley National Laboratory, Berkeley, California 94720}
-      \affil[$\dagger$]{\textit {\href{mailto:ckoerber@berkeley.edu}{ckoerber@berkeley.edu}}}
+      \affil[$\dagger$]{\textit {\href{mailto:christopher@ckoerber.com}{christopher@ckoerber.com}}}
 
   \date{\vspace{-5ex}}
 
