### 2019-12-07 63117f8
Updated margin to fit arxiv banner and updated links
