+++ doc-src/arXiv/arXiv.tex
@@ -39,8 +39,8 @@
 
 
 % --- Page layout -------------------------------------------------------------
-\usepackage[top=3.5cm, bottom=3cm, right=1.5cm, left=1.0cm,
-            headheight=2.2cm, reversemp, includemp, marginparwidth=4.5cm]{geometry}
+\usepackage[top=3.5cm, bottom=3cm, right=1.5cm, left=1.5cm,
+            headheight=2.2cm, reversemp, includemp, marginparwidth=4.0cm]{geometry}
 
 % --- Default font ------------------------------------------------------------
 % \renewcommand\familydefault{\sfdefault}
@@ -199,7 +199,7 @@
       \affil[1]{iTHEMS RIKEN, Wako, Saitama 351-0198, Japan}
       \affil[2]{Department of Physics, University of California, Berkeley, California 94720, USA}
       \affil[3]{Nuclear Science Division, Lawrence Berkeley National Laboratory, Berkeley, California 94720}
-      \affil[$\dagger$]{\textit {\href{mailto:christopher@ckoerber.com}{christopher@ckoerber.com}}}
+      \affil[$\dagger$]{\textit {\href{mailto:software@ckoerber.com}{software@ckoerber.com}}}
 
   \date{\vspace{-5ex}}
 
@@ -212,7 +212,7 @@
   %\hrule
   \sffamily\small
 
-  {\bfseries DOI:} \href{XXX}{\color{linky}{TBD}}
+  %{\bfseries DOI:} \href{XXX}{\color{linky}{TBD}}
 
   \vspace{2mm}
 
@@ -220,17 +220,16 @@
   \begin{itemize}
     \setlength\itemsep{0em}
     \item \href{https://github.com/openjournals/joss-reviews/issues/1936}{\color{linky}{Review}} \ExternalLink
-    \item \href{https://github.com/callat-qcd/espressodb}{\color{linky}{EspressoDB Repository}} \ExternalLink
-    \item \href{https://github.com/callat-qcd/lattedb}{\color{linky}{LattedDB Repository}} \ExternalLink
-    \item \href{https://ithems.lbl.gov/lattedb}{\color{linky}{LattedDB Webpage}} \ExternalLink
+    \item \href{https://github.com/callat-qcd/espressodb}{\color{linky}{EspressoDB}} \ExternalLink
+    \item \href{https://github.com/callat-qcd/lattedb}{\color{linky}{LattedDB}} \ExternalLink
   \end{itemize}
 
 
 
   \vspace{2mm}
 
-  {\bfseries Submitted:} 06 December 2019\\
-  {\bfseries Published:} XX XXX 20XX
+  {\bfseries Submitted:} 06 Dec. 2019\\
+  %{\bfseries Published:} XX XXX 20XX
 
   \vspace{2mm}
   {\bfseries Licence}\\
