### 2019-12-07 9a6d434
Changed figure location
