### 2019-12-07 4388eb5
Cleaned up root makefile
