### 2019-12-07 f050916
Cleaned up root makefile
