### 2020-01-02 #1936 @ixjlyons
Apologies for the delay. I'm able to review.
