### 2020-01-09 #1936 @ckoerber
Happy New Year!

Thank you for agreeing to referee our paper, @ixjlyons.

@gkthiruvathukal, please let us know if you need any additional information from us.
