### 2020-01-09 #2007 @whedon
[REVIEW]: EspressoDB: A scientific database for managing high-performance computing workflows
**Submitting author:** @ckoerber (<a href="http://orcid.org/0000-0002-9271-8022">Christopher Körber</a>)
**Repository:** <a href="https://github.com/callat-qcd/espressodb" target ="_blank">https://github.com/callat-qcd/espressodb</a>
**Version:** v1.1.0
