### 2020-01-20 #2007 @ixjlyons
Same here -- I've made progress but haven't updated my checklist. I should be able to finish in the next week or so as well. Thanks for the reminder.
