### 2020-01-24 #2007 @remram44
The software side looks fine to me! There are automated tests, extensive documentation including examples.

One missing item is the "community guidelines" explaining how to contribute, but that's easy to add. (I don't know if *you need one*, but it's on the JOSS checklist :wink:)

Here are my remarks looking around the documentation and code:

* there is a branch `v1.0.0` with the same name as the tag `v1.0.0`, which chokes Git sometimes (it doesn't know which one you mean, even though they point to the same commit). You should probably remove the branch (to avoid an error message when deleting it, use the GitHub UI or the command `git push origin :refs/heads/v1.0.0`)
* Git tag names are inconsistent (`0.1.0`, `0.2.1`, `v1.0.0`). You should stick with either using the prefix or not. The other tags were `< 1.0` so if you're just using the prefix from now on I don't think you absolutely have to fix the older tags.
* the `blackmagicsorcery` module stands out a bit, especially now that it just re-exports Python's standard `re` module (since 31c69150). You should probably just import `re` directly.
* the docs use both RST and MD files (but they seem to render correctly on ReadTheDocs)

Regarding the paper markup, it looks fine, although you probably don't need to monospace "Django" since it's the official name of the project, not just a package import name. "Leadership computing facilities" should be "Leading computing facilities".

-----

My main concerns are about the motivation for the project. In particular, I am not sure of the value added compared to Django, which provides a lot of the claimed features of EspressoDB, some of which being just re-exported, such as database management, signals, and automatically creating views from models (Django has the "admin" pages).

* It seems to me that it might be more useful if it was easier to integrate with existing data and processes? Creating a full Django project is heavy (project folder, app folder, migrations, ... are still there and need to be dealt with manually).
* Maybe it should have more features to keep track of experiment runs and data dependencies (e.g. which data was created from which version of which data), kind of like a workflow system? The way [data is added in the example](https://espressodb.readthedocs.io/en/latest/example/project-creation/adding-data.html#data-scripts) seems kind of dangerous, with no versioning. [This example update script](https://espressodb.readthedocs.io/en/latest/example/project-creation/dependencies.html#updating-the-run-script) has to manually check the data, counting the records to figure out if the data is up to date or not.
* It's not clear how EspressoDB helps "centralize and guarantee" data integrity from my reading of the docs. When multiple versions of the code are around with multiple people working on it, how does this code interact with a centralize database? How does the system deal with revisions of the code? Can it guarantee integrity in the presence of buggy code, which might for example attempt to delete everything?
* Also note that LatteDB is the only live example in the paper, and it doesn't let me see any data (I get a login form), though it lets me open a submission form. This doesn't seem in line with "supports [...] open-data oriented projects", although I'm sure you might have your reasons for this configuration.

There is also no related work in the paper, although data management and workflow management are big fields. Some things that come to mind:

* data management systems, such as [DVC](https://dvc.org/), which people might use to fill the same need
* workflow management systems ([Taverna](https://taverna.incubator.apache.org/), [Galaxy](https://galaxyproject.org/), [VisTrails](https://www.vistrails.org/), ...) which seem very relevant after reading your Summary section
  * **full disclosure: I worked on VisTrails**, though the project is now dormant
* data repositories: if hosting data is the primary function, adding features to an extensible data repository system naively seems like a good approach, compared to building from a generic web framework like Django (even with the added features of EspressoDB)?
