### 2020-01-26 callat-qcd/espressodb#41 @ixjlyons
[JOSS review] Installation from source requires pre-installed Django
Minor issue. The README currently lists the following steps to install EspressoDB from source for testing:

> To run the tests, clone this repo, install the dependencies

```bash
pip install .
pip install -r requirements-dev.txt
pip install -r example/my_project/requirements.txt
```

The first step fails in a clean environment because `espressodb/__init__.py` imports Django. It may also be worth noting that installing the sdist from PyPI also fails for the same reason:

```bash
pip install --no-binary espressodb espressodb
```

I see the latter as pretty much a non-issue though since you're providing a universal wheel. I think the simplest solution would be to re-order the steps so the requirements are installed first. Perhaps there's a way to defer importing Django for settings -- I'm not sure.

https://github.com/openjournals/joss-reviews/issues/2007
