### 2020-01-26 0c08661
Doc typo fixes
