### 2020-01-29 #2007 @ixjlyons
I've completed my review. Overall, EspressoDB provides a fairly rich set of features for relatively little programming effort. It seems to fulfill an important role in managing complex datasets while encouraging documentation (by making it simple to generate), which is often overlooked. The documentation of EspressoDB itself is fairly thorough and provides a straightforward path to getting started and moving on to more advanced usage. I think it fits into the scope of JOSS and needs just a few improvements in my opinion.

I ended up filing one issue regarding some development installation instructions in the README and I also made a pull request with some typo/grammar fixes in the docs. Aside from that, the following are minor issues I noticed that I didn't feel warranted actual issues against the repository:

- Links to Django docs are inconsistent, sometimes using dev docs and sometimes a specific version (e.g. v2.2 in `Usage.md`). May want to update for consistency.
- In the example/project-creation/class-interface doc, `python manage.py shell` is introduced and indicates you'll get an ipython shell, however ipython isn't in the example `requirements.txt`. Consider adding a note in the doc and/or adding ipython to the example dependencies.

Minor grammatical issues in the paper:

First line of the "Use case" section is a fragment. Consider using something like: "LatteDB, an application of ... calculations and analysis, is currently being..." or "LatteDB is an application of ..."

Also in use case section: "ultimately processed down to hundred of..." -> "hundreds of".

Consider: "...status of these files in real-time (identify corrupt..." -> "...status of these files in real-time to identify corrupt..."

I've so far left a few checklist items un-checked. Here is some rationale for these:

- Statement of need (both the documentation and paper): I think this is mostly in place, but I'm not sure I completely understand the use case(s) for EspressoDB. See below for a bit more on this.
- Community guidelines: this seems to be missing. A few sentences in the README or a contributing doc would suffice just to point out how to get help, report issues, and contribute.
- State of the field: I'm not all that familiar with related tools, but I expected some discussion of other data management solutions (e.g. Data Version Control). If it's only to mention why existing tools don't suit the needs of computational physics and such, that's fine.

Finally, I have a couple more open-ended thoughts you might consider.

There seems to be a pretty heavy reliance on users coming in with some understanding of Django to do anything nontrivial. This isn't necessarily a bad thing, but it could be stated early in the documentation that this is the case. EspressoDB (probably rightly) doesn't attempt to abstract away from Django to avoid this, so I think users should be informed up front.

A potential issue I see with this framework is that it seems to bring together system administration and data processing in a way that I'm not sure is ideal. Perhaps some explanation of *how* LatteDB (or a hypothetical example instead) is implemented could be sufficient. I'm left wondering, for example, who manages the system and who uses it to do computational work. If everyone using the system (i.e. writing data processing code) needs know Django and/or web development, the applicability of the framework may be somewhat limited. The paper and/or docs might benefit from some explanation of a reasonable workflow involving multiple team members with different roles (e.g. non-scientific administrator, computation-focused programmers, scientists pulling data to do local analyses, etc.).
