### 2020-01-29 callat-qcd/espressodb#41 @ckoerber
Hello @ixjlyons,

Thank you for pointing out this issue.

The easiest fix would indeed be updating the docs to run the `requirements` install before the `espressodb` install.

I believe it should be possible to place the `espressodb/__init__.py` Django import in the provided `init` function. This would delay the import until `init()` is actually called and thus allows to run a `no-binary` install. I have to run some checks&mdash;if this works out and does not affect other functionalities, I will include the delayed import in the next version.
