### 2020-01-30 callat-qcd/espressodb#43 @ckoerber
Paper & Documentation improvements
Updates related to the JOSS review process raised by @remram44 and @ixjlyons.

## Documentation improvements

Improvements to be made for the documentation:

- [x] Consistent links to Django docs (dev version)
- [x] Remarks about which shell is used (espressodb/blob/master/doc-src/example/project-creation/class-interface.md)
- [x] Readme: how much knowledge of Django do you need to do what?
- [x] Add a statement comparing EspressoDB to, e.g., [DVC](https://github.com/iterative/dvc), [Taverna](https://taverna.incubator.apache.org/), [Galaxy](https://galaxyproject.org), [VisTrails](https://www.vistrails.org/index.php/Main_Page), ..
- [x] Add versioning statement in advanced features of EspressoDB
- [x] Statement about access levels and data loss.

## Paper improvements

Improvements to be made for the paper:

- [x] First line of the "Use case" section is a fragment. Consider using something like: "LatteDB, an application of ... calculations and analysis, is currently being..." or "LatteDB is an application of ..."
- [x] Also in use case section: "ultimately processed down to hundred of..." -> "hundreds of".
- [x] Consider: "...status of these files in real-time (identify corrupt..." -> "...status of these files in real-time to identify corrupt..."
- [x] Monospace Django -> Django
- [x] Add a statement comparing EspressoDB to, e.g., [DVC](https://github.com/iterative/dvc), [Taverna](https://taverna.incubator.apache.org/), [Galaxy](https://galaxyproject.org), [VisTrails](https://www.vistrails.org/index.php/Main_Page), ...

https://github.com/openjournals/joss-reviews/issues/2007
