### 2020-01-30 callat-qcd/espressodb#44 @ckoerber
Community guidelines
We need to add community guidelines, e.g., how to file issues and how to contribute in a prominent documentation place. See also https://opensource.guide/starting-a-project/#writing-your-contributing-guidelines.

Update related to the JOSS review process raised by @remram44 and @ixjlyons.

https://github.com/openjournals/joss-reviews/issues/2007
