### 2020-01-30 callat-qcd/espressodb#46 @ckoerber
Version checks between local and remote projects
In the scenario where a user communicates with a central database running different table definitions (different migration histories), it might happen that intermediate objects are inserted before the user code crashes.

This can be prevented by running checks on the import of EspressoDB (the `init()` function).
Things to cross check are:

(1) The version of EspressoDB
(2) The version of the EspressoDB project if applicable
(3) The migration history

The execution of checks should be optional to allow rapid development--nevertheless, this information should be on a prominent doc page.

* The migration history check can be realized by the default mechanics of [`makemigrations`](https://github.com/django/django/blob/5d654e1e7104d2ce86ec1b9fe52865a7dca4b4be/django/core/management/commands/makemigrations.py#L87).
* EspressoDB and EspressoDB project versions must be inserted in the database when migrations are applied. This can be done by adding new default version tables and overloading the `manage.py migrate` command. On import (and if required) EspressoDB tries to read the most recent information and fails if versions disagree.

Update related to the JOSS review process raised by @remram44.

https://github.com/openjournals/joss-reviews/issues/2007
