### 2020-01-30 #2007 @ckoerber
Hello @gkthiruvathukal, @ixjlyons, and @remram44,

Thank you for your time and the feedback you have provided. We believe that your comments help to improve EsspressoDB. To keep updates transparent, we filed issues for suggested changes and intend to merge them into the new version `v1.1.0`. The filed issues are collected in a new [project on EsspressoDB](https://github.com/callat-qcd/espressodb/projects/6). We believe that we should be able to finalize the new features by the end of next week.

In the next days, we will also address the points you have made in more detail.

General statements regarding the open checkboxes:

* [We will add community guidelines in the next release](https://github.com/callat-qcd/espressodb/issues/44).
* [We will add a statement addressing the state of the field in both documentation and paper](https://github.com/callat-qcd/espressodb/issues/43).
* The publication of EspressoDB goes hand-in-hand with LatteDB. Because LatteDB is domain-specific (Lattice Quantum Chromodynamics), we decided to submit EspressoDB as it might be more beneficial to a broader audience. In our paper, we rather address the need for EspressoDB from our science domain by pointing out the challenges of Lattice computations.

For example

> ... CalLat creates petabytes of temporary files that are written to the scratch file system, used for subsequent computations and ultimately processed down to hundred of tera-bytes that are saved for analysis. It is essential to track the status of these files in real-time (identify corrupt, missing, or purgeable files).

To address the statement of need, should we be more explicit about how LatteDB (and thus EspressoDB) helps to track files or rather approach it from a more general point of view?
