### 2020-01-30 #2007 @gkthiruvathukal
@gkthiruvathukal Just letting @ckoerber and all know that I'm keeping an eye on the thread. It sounds like the feedback from @ixjlyons, and @remram44 has been well received and there is a plan to work on the issues raised during review.

After the issues are addressed, I'll have the reviewers take another look.

I'd like to ask for everyone's help here. I have no reason to doubt that all authors of the software are represented on the paper submission, but can each of you (authors and reviewers) please confirm for me?

"Does the full list of paper authors seem appropriate and complete?"

Yes, I know the checkbox has been checked but am asking you to check once more. I am in the middle of dealing with another submission (not edited by me) where the answer is "no" so I am now checking every one of my editorial assignments to make sure there are no authors--or potential authors--who are not listed.  If you can do a brief follow-up, this would be much appreciated.
