### 2020-01-30 #2007 @remram44
I confirm that @ckoerber made major contributions to the software, and that the list of authors match the major contributors of the software.

It seems that @cchang5 goes by "Jason Chang" on his GitHub profile and "Chia Cheng Chang" on the paper, but it looks like this is intended.
