### 2020-01-30 #2007 @cchang5
@gkthiruvathukal
I'd like to ask for everyone's help here. I have no reason to doubt that all authors of the software are represented on the paper submission, but can each of you (authors and reviewers) please confirm for me?

I am confirming that I am an author of the software.
