### 2020-01-30 #2007 @ckoerber
I can confirm that @cchang5, @walkloud, and I are the authors of EspressoDB (and LatteDB).
