### 2020-01-30 #2007 @ixjlyons
Confirming I've re-checked the author list against the contributors based on git history.
