### 2020-02-03 9df504e
Moved Django imports into init()
This addresses #41
