### 2020-02-03 callat-qcd/espressodb#41 @ckoerber
The commit 9df504e should resolve this issue (no doc update needed). This should also allow the `no-binary` install which will be tested once the new version is ready for shipping.
