### 2020-02-04 12ef307
All workflow tests are now executed by example/Makefile
