+++ Makefile
@@ -18,12 +18,5 @@ docs:
 	make -C doc-src html
 
 .PHONY: test
-test: test_my_project test_m2m
-
-.PHONY: test_my_project
-test_my_project:
-	cd example/my_project && pytest
-
-.PHONY: test_m2m
-test_m2m:
-	cd example/espressodb_tests && pytest
+test:
+	make -C example/ test
