### 2020-02-04 2d4704a
Added first draft of CONTRIBUTING
