### 2020-02-04 a3ef9d5
Added contributing link to readme
