+++ README.md
@@ -84,6 +84,9 @@ The core module of LatteDB, EspressoDB, is trying to address those challenges.
 * [@ckoerber](https://www.ckoerber.com) (Feel free to contact me for questions)
 * [@cchang5](https://github.com/cchang5)
 
+## Contributing
+Thanks for your interest in contributing! There are many ways to contribute to this project. 
+[Get started here](CONTRIBUTING.md).
 
 ## License
 
