### 2020-02-04 69b3e45
Added contributing to docs.
This addresses #44.
