+++ doc-src/index.rst
@@ -18,3 +18,4 @@ Welcome to the documentation of `EspressoDB` -- an open source Python package fo
    example/index
    features/index
    api/index
+   CONTRIBUTING
