### 2020-02-05 19ac1c8
Fixed docstring
