### 2020-02-05 a441aec
Added pre_save functionality to Base class
