### 2020-02-05 1066b3a
Added pre_save unittests
