### 2020-02-05 55447a0
Added documentation about pre-save and code revision information.
This and the previous commits address #45
