+++ doc-src/features/index.rst
@@ -11,4 +11,5 @@ In this section we present
     :caption: Contents:
 
     consistency-checks
+    pre-save
     model-specializations
