### 2020-02-05 ceef601
Added remarks for disabling pre-save
