+++ doc-src/example/project-creation/dependencies.md
@@ -100,7 +100,7 @@ Running this script will compute all the eigenvalues.
 <div class="admonition warning">
 <p class="admonition-title">Note</p>
 <p>
-    EspressoDB automatically stores `user` and `date` information for each entry.
+    EspressoDB automatically stores `user` and `last_modified` information for each entry.
     In some cases it might be desirable to also store code revision information.
     This can be done by specifying the `tag` column.
     See also [the `pre_save` functionality.](../../features/pre-save.md).
