### 2020-02-05 e8a331f
Added checks which test the state of models, migrations and the database
This addresses #46.
Options for enabling checks and tests need to be added.
