+++ espressodb/management/checks/__init__.py
@@ -1,0 +1,4 @@
+"""Checks for the state of local migration files and models compared to database
+
+* showmigrations lists local applied and unapplied migrations
+"""
