### 2020-02-06 f5413b4
Added runtime checks for migration state on espressodb init
