### 2020-02-06 f30da92
Added first test to check if run checks is called as expected
