### 2020-02-06 ca2e531
Moved tests to tests directory
