+++ example/README.md
@@ -1,0 +1,3 @@
+# What to find here
+
+The `my_project` directory contains an example EspressoDB project.
