+++ example/espressodb_tests/db-config.yaml
@@ -1,2 +1,0 @@
-ENGINE: django.db.backends.sqlite3
-NAME: espressodb-tests-db.sqlite
