+++ .github/workflows/run-tests.yml
@@ -37,19 +37,19 @@ jobs:
               run: pytest
 
             - name: Run m2m tests
-              working-directory: ./example/espressodb_tests/
+              working-directory: ./tests/espressodb_tests/
               run: pytest
 
             - name: Run create new project
-              working-directory: ./example/
+              working-directory: ./tests/
               run: espressodb startproject espressodb_new_test
 
             - name: Test new project
-              working-directory: ./example/espressodb_new_test/
+              working-directory: ./tests/espressodb_new_test/
               run: python manage.py test
 
             - name: Run start app
-              working-directory: ./example/espressodb_new_test/
+              working-directory: ./tests/espressodb_new_test/
               run: |
                 python manage.py startapp new_app
                 echo "" >> espressodb_new_test/new_app/models.py
@@ -59,5 +59,5 @@ jobs:
                 python manage.py makemigrations
 
             - name: Test new project app
-              working-directory: ./example/espressodb_new_test/
+              working-directory: ./tests/espressodb_new_test/
               run: python manage.py test
