+++ tests/migration_tests/manage.py
@@ -1,0 +1,25 @@
+#!/usr/bin/env python
+"""Django's command-line utility for administrative tasks."""
+import os
+import sys
+
+
+def main():
+    os.environ.setdefault(
+        "DJANGO_SETTINGS_MODULE", "migration_tests.config.settings"
+    )
+    os.environ["ESPRESSODB_INIT_CHECKS"] = "0"  # Do not run checks for manage.py
+
+    try:
+        from django.core.management import execute_from_command_line
+    except ImportError as exc:
+        raise ImportError(
+            "Couldn't import Django. Are you sure it's installed and "
+            "available on your PYTHONPATH environment variable? Did you "
+            "forget to activate a virtual environment?"
+        ) from exc
+    execute_from_command_line(sys.argv)
+
+
+if __name__ == "__main__":
+    main()
