+++ tests/migration_tests/migration_tests/config/settings.py
@@ -1,0 +1,174 @@
+"""
+Django settings for migration_tests project.
+
+Generated by 'django-admin startproject' using Django 2.2.8.
+
+For more information on this file, see
+https://docs.djangoproject.com/en/2.2/topics/settings/
+
+For the full list of settings and their values, see
+https://docs.djangoproject.com/en/2.2/ref/settings/
+
+This file has been adjusted to espressodb settings.
+"""
+import os
+
+
+from espressodb.management.utilities.files import get_project_settings
+from espressodb.management.utilities.files import get_db_config
+from espressodb.management.utilities.files import ESPRESSO_DB_ROOT
+
+# Build paths inside the project like this: os.path.join(BASE_DIR, ...)
+BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
+ROOT_DIR = os.path.dirname(BASE_DIR)
+
+_SETTINGS = get_project_settings(ROOT_DIR)
+SECRET_KEY = _SETTINGS.get("SECRET_KEY")
+PROJECT_APPS = _SETTINGS.get("PROJECT_APPS", [])
+ALLOWED_HOSTS = _SETTINGS.get("ALLOWED_HOSTS", [])
+DEBUG = _SETTINGS.get("DEBUG", False)
+
+
+# Quick-start development settings - unsuitable for production
+# See https://docs.djangoproject.com/en/2.2/howto/deployment/checklist/
+
+
+# Application definition
+
+INSTALLED_APPS = (
+    PROJECT_APPS
+    + [
+        "espressodb.base",
+        "espressodb.documentation",
+        "espressodb.management",
+        "espressodb.notifications",
+    ]
+    + ["bootstrap4", "widget_tweaks"]
+    + [
+        "django.contrib.admin",
+        "django.contrib.auth",
+        "django.contrib.contenttypes",
+        "django.contrib.sessions",
+        "django.contrib.messages",
+        "django.contrib.staticfiles",
+    ]
+)
+
+MIDDLEWARE = [
+    "django.middleware.security.SecurityMiddleware",
+    "django.contrib.sessions.middleware.SessionMiddleware",
+    "django.middleware.common.CommonMiddleware",
+    "django.middleware.csrf.CsrfViewMiddleware",
+    "django.contrib.auth.middleware.AuthenticationMiddleware",
+    "django.contrib.messages.middleware.MessageMiddleware",
+    "django.middleware.clickjacking.XFrameOptionsMiddleware",
+]
+
+ROOT_URLCONF = "migration_tests.config.urls"
+
+TEMPLATES = [
+    {
+        "BACKEND": "django.template.backends.django.DjangoTemplates",
+        "DIRS": [
+            os.path.join(ESPRESSO_DB_ROOT, "espressodb", "base", "templates"),
+            os.path.join(ESPRESSO_DB_ROOT, "espressodb", "documentation", "templates"),
+        ],
+        "APP_DIRS": True,
+        "OPTIONS": {
+            "context_processors": [
+                "django.template.context_processors.debug",
+                "django.template.context_processors.request",
+                "django.contrib.auth.context_processors.auth",
+                "django.contrib.messages.context_processors.messages",
+            ]
+        },
+    }
+]
+
+for app in PROJECT_APPS[::-1]:
+    _template_dir = os.path.join(ROOT_DIR, app.replace(".", os.sep), "templates")
[…]
