+++ tests/migration_tests/README.md
@@ -1,0 +1,4 @@
+# Migration tests
+
+This app tests whether migration checks run as expected.
+To make this work, this test project contains three apps with complete and incomplete migration status to trigger exceptions.
