+++ tests/migration_tests/settings.yaml
@@ -1,6 +1,4 @@
 SECRET_KEY: "h6$^j)0e8g31^e-2)-%4s*&e8b9stbr*j!r$*20tp5jqs))-j@"
-PROJECT_APPS:
-    - migration_tests.m2mtests
-    - espressodb_tests.pre_save_tests
+PROJECT_APPS: []
 ALLOWED_HOSTS: ["testserver", "localhost", "127.0.0.1"]
 DEBUG: True
