+++ tests/migration_tests/migration_tests/app1/models.py
@@ -1,7 +1,5 @@
-"""Models of app1
+"""This apps models migrations and database are all fine.
 """
-
-# Note: if you want your models to use espressodb features, they must inherit from Base
 
 from django.db import models
 from espressodb.base.models import Base
