+++ tests/migration_tests/manage.py
@@ -2,11 +2,37 @@
 """Django's command-line utility for administrative tasks."""
 import os
 import sys
+from espressodb import init as _init
+
+APP_ID = None
 
 
 def main():
-    os.environ.setdefault(
-        "DJANGO_SETTINGS_MODULE", "migration_tests.config.settings"
+    """Initializes the module as standalone package without settings file.
+    """
+
+    CWD = os.path.dirname(os.path.abspath(os.path.dirname(__file__)))
+
+    PROJECT_APPS = [f"migration_tests.app{APP_ID}"] if APP_ID is not None else []
+    _init(
+        DATABASES={
+            "default": {
+                "ENGINE": "django.db.backends.sqlite3",
+                "NAME": f"migration-tests-db-state"
+                + (f"-{APP_ID}" if APP_ID is not None else "")
+                + ".sqlite",
+            }
+        },
+        INSTALLED_APPS=PROJECT_APPS
+        + [
+            "espressodb.base",
+            "espressodb.management",
+            "django.contrib.auth",
+            "django.contrib.contenttypes",
+        ],
+        PROJECT_APPS=PROJECT_APPS,
+        ROOT_DIR=CWD,
+        PROJECT_NAME="migration_tests",
     )
     os.environ["ESPRESSODB_INIT_CHECKS"] = "0"  # Do not run checks for manage.py
 
