### 2020-02-06 c985f83
Started to write migration state tests
