+++ tests/migration_tests/Makefile
@@ -1,0 +1,18 @@
+.PHONY: test
+test: test_state1 test_state2 test_state3 test_state4
+
+.PHONY: test_state1
+test_state1:
+	pytest tests/test_state1.py
+
+.PHONY: test_state2
+test_state2:
+	pytest tests/test_state2.py
+
+.PHONY: test_state3
+test_state3:
+	pytest tests/test_state3.py
+
+.PHONY: test_state4
+test_state4:
+	pytest tests/test_state4.py
