+++ tests/migration_tests/pytest.ini
@@ -1,5 +1,5 @@
 [pytest]
-python_files = tests.py test_*.py *_tests.py
+python_files = test_*.py *_tests.py
 addopts = -v -p no:django
 env =
     ESPRESSODB_INIT_CHECKS=0
