### 2020-02-06 f3f7cd7
Finalized migration state tests
