### 2020-02-06 callat-qcd/espressodb#46 @ckoerber
For this issue, the code development takes place on the [`feature-init-checks`](https://github.com/callat-qcd/espressodb/tree/feature-init-checks) branch.
