### 2020-02-06 6b29c2a
Makefile test runs for migration checks
