### 2020-02-06 854fe2c
Added migration tests to github workflow
