+++ .github/workflows/run-tests.yml
@@ -36,9 +36,17 @@ jobs:
               working-directory: ./example/my_project/
               run: pytest
 
-            - name: Run m2m tests
+            - name: Run espressodb tests
               working-directory: ./tests/espressodb_tests/
               run: pytest
+
+            - name: Run migration check tests
+              working-directory: ./tests/migration_tests/
+              run: |
+                pytest tests/test_state1.py
+                pytest tests/test_state2.py
+                pytest tests/test_state3.py
+                pytest tests/test_state4.py
 
             - name: Run create new project
               working-directory: ./tests/
