### 2020-02-06 02e8302
Added description of how to run tests
