### 2020-02-07 718544c
Added docs for new init checks
