### 2020-02-07 8e299e9
Doc string updates
