+++ tests/espressodb_tests/espressodb_tests/__init__.py
@@ -19,7 +19,7 @@ def _init():
             msg = "Failed to import EspressoDB project!\n\n"
             msg += str(error)
             msg += "\n\nYou are seeing this error because,"
-            msg += " on initialization, EspressoDB runs cross checks."
+            msg += " on initialization, EspressoDB runs cross-checks."
             msg += " If you want to disable this behavior, set the"
             msg += " environment variable `ESPRESSODB_INIT_CHECKS=0`."
 
