+++ espressodb/__init__.py
@@ -2,7 +2,7 @@
 """Initializes minimal settings to launch EspressoDB
 """
 
-__version__ = "1.0.0"
+__version__ = "1.1.0"
 
 DEFAULT_OPTIONS = {
     "DEBUG": True,
