### 2020-02-07 316f9cd
Updated django docs 2.2 links to dev links
