+++ tests/espressodb_tests/espressodb_tests/config/urls.py
@@ -1,7 +1,7 @@
 """espressodb_tests URL Configuration
 
 The `urlpatterns` list routes URLs to views. For more information please see:
-    https://docs.djangoproject.com/en/2.2/topics/http/urls/
+    https://docs.djangoproject.com/en/dev/topics/http/urls/
 Examples:
 Function views
     1. Add an import:  from my_app import views
