+++ doc-src/example/project-creation/views.md
@@ -21,7 +21,7 @@ Here is an example of how to overwrite the index page.
     <h5>This page is generated by the template <code>my_project/hamiltonian/templates/index.html</code> view</h2>
     <p>
         See also the
-        <a href="https://docs.djangoproject.com/en/2.2/ref/templates/language/">
+        <a href="https://docs.djangoproject.com/en/dev/ref/templates/language/">
             the django template documentation
         </a>
         how templates work.
