+++ paper.md
@@ -42,21 +42,21 @@ A typical scientific computing workflow includes:
 4. Post-processing these data files;
 5. Performing data analysis on output.
 
-[``EspressoDB``](https://github.com/callat-qcd/espressodb/) is a programmatic object-relational data management framework implemented in Python and based on the [Django web framework](https://www.djangoproject.com).
-``EspressoDB`` was developed to streamline data management workflows, centralize and guarantee data integrity, while providing domain flexibility and ease of use.
+[EspressoDB](https://github.com/callat-qcd/espressodb/) is a programmatic object-relational data management framework implemented in Python and based on the [Django web framework](https://www.djangoproject.com).
+EspressoDB was developed to streamline data management workflows, centralize and guarantee data integrity, while providing domain flexibility and ease of use.
 
-The framework provided by ``EspressoDB`` aims to support the ever-increasing complexity of workflows of scientific computing at leadership computing facilities (LCFs), with the goal of reducing the amount of human time required to manage the jobs, thus giving scientists more time to focus on science.
+The framework provided by EspressoDB aims to support the ever-increasing complexity of workflows of scientific computing at leadership computing facilities (LCFs), with the goal of reducing the amount of human time required to manage the jobs, thus giving scientists more time to focus on science.
 
 # Features
 Data integrity is important to scientific projects and becomes more challenging the larger the project.
 In general, a ``SQL`` framework type-checks data before writing to the database and controls dependencies and relations between different tables to ensure internal consistency.
- ``EspressoDB`` allows additional user-defined constraints not supported by ``SQL`` (*e.g.* unique constraints using information across related tables).
-Once the user has specified a set of conditions that entries have to fulfill for each table, ``EspressoDB`` runs these cross-checks for new data before inserting them in the database.
+ EspressoDB allows additional user-defined constraints not supported by ``SQL`` (*e.g.* unique constraints using information across related tables).
+Once the user has specified a set of conditions that entries have to fulfill for each table, EspressoDB runs these cross-checks for new data before inserting them in the database.
 
-``EspressoDB`` also supports collaborative and open-data oriented projects by leveraging and extending Django's web hosting component.
+EspressoDB also supports collaborative and open-data oriented projects by leveraging and extending Django's web hosting component.
 In addition to providing a centralized data platform, it is possible to spawn customized web pages which can be hosted locally or on the world wide web[^1].
-``EspressoDB`` simplifies creating projects by providing default Django configurations that set up, for example, connections to the database and webpages to view associated tables.
-For example, with the default setting, ``EspressoDB`` spawns:
+EspressoDB simplifies creating projects by providing default Django configurations that set up, for example, connections to the database and webpages to view associated tables.
+For example, with the default setting, EspressoDB spawns:
 
 * Documentation views of implemented tables;
 * A project-wide notification system;
@@ -72,9 +72,9 @@ More details, usage instructions, and examples are documented at [espressodb.rea
 
 
 # Use case
-[``LatteDB``](https://github.com/callat-qcd/lattedb/) is an application of ``EspressoDB`` that is specialized to contain table definitions for lattice quantum chromodynamics (LQCD) calculations and analysis.
[…]
