+++ doc-src/arXiv/use-case.md
@@ -1,5 +1,5 @@
 # Django overview
-``EspressoDB`` utilizes Python's Django Object-Relational Mapping (ORM) framework.
+EspressoDB utilizes Python's Django Object-Relational Mapping (ORM) framework.
 Tables correspond to Python classes, rows correspond to instances and columns correspond to attributes of the instances.
 Thus it is possible to filter for objects by their attributes or generate summary tables (``pandas.DataFrame``) within one line of code.
 Furthermore, using an ORM allows one to have the same interface independent of the backend.
@@ -7,7 +7,7 @@ It is possible to store data in a file based `SQLite` solution, or use more scal
 
 Django is part of many open-source projects and thus comes with extensive documentation.
 Additionally, Django is scalable, comes with reliable tests and vast community support which manifests in the fact that it is commonly  used in large scale projects (BitBucket, Instagram, Mozilla, NASA and many more).
-One guiding principle of ``EspressoDB`` is to not "re-invent the wheel" but instead leverage the support coming from Django.
+One guiding principle of EspressoDB is to not "re-invent the wheel" but instead leverage the support coming from Django.
 As a result, one can easily incorporate many of Django's extensions and find solutions to technical questions online.
 
 # Lattice QCD use case
@@ -36,19 +36,19 @@ As a concrete example, we consider the nucleon elastic form factor project being
 7. Process the data to prepare it for analysis and the extraction of physics.
 8. Repeat for multiple source sets (of 8) if more statistics are required.
 
-Each of the steps above leads to the generation of many large (multi-gigabyte) data files, most of which are not saved.  ``LatteDB`` is used to track the status of these data files to know if they need to be created or if the next step can proceed.  The final data production step, 6, leads to our final data file that need further processing prior to our final analysis and saving of the files.
+Each of the steps above leads to the generation of many large (multi-gigabyte) data files, most of which are not saved.  LatteDB is used to track the status of these data files to know if they need to be created or if the next step can proceed.  The final data production step, 6, leads to our final data file that need further processing prior to our final analysis and saving of the files.
 
-Inheriting the functionality of ``EspressoDB``, ``LatteDB`` has the flexibility to faithfully reproduce a one-to-one mapping between the above computational workflow to database tables.
+Inheriting the functionality of EspressoDB, LatteDB has the flexibility to faithfully reproduce a one-to-one mapping between the above computational workflow to database tables.
 For example, in step 1, the table of gauge configurations is defined such that every row in the table specifies a single gauge configuration.
 This reflects how on disk, we have thousands of files, each containing a snapshot of the QCD vacuum, and as such, every file, and every subsequent file as a result of the gauge configuration (*e.g.* propagators or correlation functions in steps 2 through 6) can also be tracked individually.
 However, at the end of the calculation, an observable is only well defined with an ensemble of gauge configurations.
-``LatteDB`` allows one to define an ensemble table, with a ``Django ManyToMany`` data type which may be interpreted as a single column containing a list of references (foreign keys) to the table of gauge configurations.
+LatteDB allows one to define an ensemble table, with a ``Django ManyToMany`` data type which may be interpreted as a single column containing a list of references (foreign keys) to the table of gauge configurations.
 In ``SQL``, a list of foreign keys is not a fundamental data type that is supported, and is only made possible with Django.
 However, even with Django, unique constraints can not be implemented on such a column.
-With ``LatteDB``, we make it possible to define a unique constraint, which for this example, prohibits the user from inserting the exact same list of gauge configurations in the ensemble table more than once.
-Users are encouraged to consult the documentation of ``EspressoDB`` and examples in ``LatteDB`` for more information.
+With LatteDB, we make it possible to define a unique constraint, which for this example, prohibits the user from inserting the exact same list of gauge configurations in the ensemble table more than once.
[…]
