+++ paper.md
@@ -72,7 +72,7 @@ More details, usage instructions, and examples are documented at [espressodb.rea
 
 
 # Use case
-[LatteDB](https://github.com/callat-qcd/lattedb/) is an application of EspressoDB that is specialized to contain table definitions for lattice quantum chromodynamics (LQCD) calculations and analysis.
+[LatteDB](https://github.com/callat-qcd/lattedb/), an application of EspressoDB, contains table definitions for lattice quantum chromodynamics (LQCD) calculations and analysis.
 LatteDB is currently being used by the [CalLat Collaboration](https://a51.lbl.gov/~callat/webhome/) in their computations on Summit at the Oak Ridge Leadership Computing Facility ([OLCF](https://www.olcf.ornl.gov)) through DOE INCITE Allocations [@incite:2019; @incite:2020].
 The website generated by LatteDB used by CalLat can be found at [https://ithems.lbl.gov/lattedb/](https://ithems.lbl.gov/lattedb/). A precursor to EspressoDB and LatteDB was used to support a series of LQCD projects [@Nicholson:2018mwc; @Chang:2018uxx].
 
@@ -82,8 +82,8 @@ Summit at OLCF is disruptively fast compared to previous generations of leadersh
 LCFs prohibit the submission of millions of small tasks to their supercomputers (clogged queues, overtaxed service nodes, etc.).  It is imperative to have a task manager capable of bundling many tasks into large jobs while distributing the work to various components of the heterogeneous nodes;
 
 2. _Dependent task generation and data processing_:
-As an example, CalLat creates petabytes of temporary files that are written to the scratch file system, used for subsequent computations and ultimately processed down to hundred of tera-bytes that are saved for analysis.
-It is essential to track the status of these files in real-time (identify corrupt, missing, or purgeable files).
+As an example, CalLat creates petabytes of temporary files that are written to the scratch file system, used for subsequent computations and ultimately processed down to hundreds of tera-bytes that are saved for analysis.
+It is essential to track the status of these files in real-time to identify corrupt, missing, or purgeable files.
 
 Members of CalLat are addressing issue 1 through the creation of job management software, [METAQ](https://github.com/evanberkowitz/metaq) [@Berkowitz:2017vcp], and ``MPI_JM`` [@Berkowitz:2018gqe; @Berkowitz:2017xna].
 LatteDB is designed to address the second issue.
