### 2020-02-07 3d36dc2
Added faq
This addresses #43
