+++ doc-src/index.rst
@@ -16,6 +16,7 @@ Welcome to the documentation of `EspressoDB` -- an open source Python package fo
    README
    Usage
    example/index
+   faq
    features/index
    api/index
    CONTRIBUTING
