### 2020-02-08 66eaa72
Added statement in the introduction
