### 2020-02-08 1d34aa0
Added team size statements
