### 2020-02-08 b032f04
Added state of the field to README
