### 2020-02-10 callat-qcd/espressodb#45 @ckoerber
This feature, docs and test have been implemented in the commit range https://github.com/callat-qcd/espressodb/compare/19ac1c8..ceef601
