### 2020-02-13 7db672a
adding description from reply-02 to paper LatteDB example
