### 2020-02-13 58b6269
Update paper.md
Updates to intro and use case

* Updated the phrase how EspressoDB is different from other Python ORMs to highlight the uniqueness and consistency checks.
* Grouped both Use Case points to avoid redundancy
