+++ paper.md
@@ -46,8 +46,8 @@ A typical scientific computing workflow includes:
 [EspressoDB](https://github.com/callat-qcd/espressodb/) is a programmatic object-relational mapping (ORM) data management framework implemented in Python and based on the [Django web framework](https://www.djangoproject.com).
 EspressoDB was developed to streamline data management, centralize and promote data integrity, while providing domain flexibility and ease of use.
 It is designed to directly integrate in utilized software to allow dynamical access to vast amount of relational data at runtime.
-Compared to existing ORM frameworks like [SQLAlchemy](https://www.sqlalchemy.org) or Django itself, EspressoDB lowers the barrier of access by simplifying the project setup.
-In contrast to software like [DVC](https://github.com/iterative/dvc), [Taverna](https://taverna.incubator.apache.org) or [VisTrails](https://www.vistrails.org/index.php/Main_Page) which describe the workflow of computations, EspressoDB rather specifies the data itself and can thus be used in a complementary spirit.
+Compared to existing ORM frameworks like [SQLAlchemy](https://www.sqlalchemy.org) or Django itself, EspressoDB lowers the barrier of access by simplifying the project setup and provides further features to satisfy uniqueness and consistency over multiple data dependencies.
+In contrast to software like [DVC](https://github.com/iterative/dvc), [Taverna](https://taverna.incubator.apache.org) or [VisTrails](https://www.vistrails.org/index.php/Main_Page) which describe the workflow of computations, EspressoDB rather interacts with data itself and thus can be used in a complementary spirit.
 
 The framework provided by EspressoDB aims to support the ever-increasing complexity of workflows of scientific computing at leadership computing facilities (LCFs), with the goal of reducing the amount of human time required to manage the jobs, thus giving scientists more time to focus on science.
 
@@ -84,18 +84,21 @@ The website generated by LatteDB used by CalLat can be found at [https://ithems.
 
 Summit at OLCF is disruptively fast compared to previous generations of leadership-class computers.  There are two challenges which are both critical to address for near-exascale computers such as Summit, which will become more important in the exascale era:
 
-1. _Efficient bundling and management of independent tasks_:
-LCFs prohibit the submission of millions of small tasks to their supercomputers (clogged queues, overtaxed service nodes, etc.).  It is imperative to have a task manager capable of bundling many tasks into large jobs while distributing the work to various components of the heterogeneous nodes;
+1. _Dependent task generation and data processing_
+2. _Efficient bundling and management of intdependent tasks_
 
-2. _Dependent task generation and data processing_:
 The computational model for lattice QCD computations can be organized as a directed multi-graph:
-a single complete calculation requires tens-of-thousands to millions of independent MPI tasks to be completed and these tasks, while independent, have nested and chained interdependencies (the output of some tasks are part of the input for other tasks).
-In order to extrapolate the results of these computations to extract final answers, several such complete computations must be performed.
-In order to manage such computations on Leadership Class computing Facilities requires these tasks to be bundled together and run as a much smaller number of large jobs and to keep the nodes from going idle, the jobs must be backfilled while running with the next set of available tasks (item 1 above).  Keeping track of the tasks, optimizing the order of tasks and ensuring no work is repeated requires a task manger that understands all these dependencies and the uniqueness of each task.  Software to track and manage such a computational model at runtime, which does not require in-depth knowledge of managing databases, does not currently exist.  As a specific example, CalLat creates petabytes of temporary files that are written to the scratch file system, used for subsequent computations and ultimately processed down to hundreds of tera-bytes that are saved for analysis.
[…]
