### 2020-02-13 171e909
Update paper.md
Removed code tags from SQL
