### 2020-02-13 callat-qcd/espressodb#49 @ckoerber
[JOSS Review] Reply to referee @ixjlyons
Thank you for your review @ixjlyons,

> I ended up filing one issue regarding some development installation instructions in the README and I also made a pull request with some typo/grammar fixes in the docs.

The PR is already merged in and we have updated EspressoDB to resolve [the issue you have filed](https://github.com/callat-qcd/espressodb/issues/41). This will be merged in with the next release.

# Paper & doc updates

We have added your doc and paper fixes to the [list of documentation and paper updates](https://github.com/callat-qcd/espressodb/issues/43).

> Links to Django docs are inconsistent, sometimes using dev docs and sometimes a specific version (e.g. v2.2 in Usage.md). May want to update for consistency.

Thank you for pointing this out. Unless breaking changes in Django are introduced, it is probably best to link against the dev docs to be consistent with the most recent version (e.g. the link will not change and they display a banner that reminds you to compare versions).

> In the example/project-creation/class-interface doc, python manage.py shell is introduced and indicates you'll get an ipython shell...

Good catch. The Django page specifies "By default, Django will use IPython or bpython if either is installed". We have added a remark in the `v1.1.0` of the EsspressoDB docs.

# General remarks

> Community guidelines: this seems to be missing. A few sentences in the README or a contributing doc would suffice just to point out how to get help, report issues, and contribute.

[Community guidelines have been added](https://github.com/callat-qcd/espressodb/issues/44).

> **Statement of need (both the documentation and paper):** I think this is mostly in place, but I'm not sure I completely understand the use case(s) for EspressoDB. See below for a bit more on this.

> **State of the field:** I'm not all that familiar with related tools, but I expected some discussion of other data management solutions (e.g. Data Version Control). If it's only to mention why existing tools don't suit the needs of computational physics and such, that's fine.

The computational model we have can be organized as a directed multi-graph: a single complete calculation requires tens-of-thousands to millions of independent MPI tasks to be completed and these tasks, while independent, have nested and chained interdependencies (the output of some tasks are part of the input for other tasks). In order to extrapolate the results of these computations to extract final answers, several such complete computations must be performed. In order to manage such computations on Leadership Class computing Facilities requires these tasks to be bundled together and run as a much smaller number of large jobs and to keep the nodes from going idle, the jobs must be backfilled while running with the next set of available tasks.  Keeping track of the tasks, optimizing the order of tasks and ensuring no work is repeated requires a task manger that understands all these dependencies and the uniqueness of each task.  Software to track and manage such a computational model at runtime, which does not require in-depth knowledge of managing databases, does not currently exist.  Consulting with our colleagues from other fields who also use Leadership Class computing Facilities, we know others would find this useful as well.  Therefore, we created EspressoDB to provide an easier interface for users to set up and maintain such a database and we created LatteDB as a specific use case to provide a real-time working example for other groups to get ideas from. Therefore, the state of the field is rather addressed by comparing EspressoDB to Object-Relational Mapping (ORM) frameworks. Since our codebase to manage and submit computations is in Python, the related tools are either SQLAlchemy or Django itself. We have promoted these comparisons in the paper and on the doc page (README).

> There seems to be a pretty heavy reliance on users coming in with some understanding of Django to do anything nontrivial. This isn't necessarily a bad thing, but it could be stated early in the documentation that this is the case.

This software also provides the benefit of lowering the technical barrier to interacting with the database.  We have in mind the following scenario: At least one user in a collaboration is familiar with databases and Python (with none to moderate familiarity with Django). Implementing a database to help manage the jobs with Django or other SQL frameworks is a sufficiently high impediment that alternative solutions are desired. For example, asking everyone to learn SQL is historically too high of a hurdle for successful collaboration-wide adoption. The intention is that EspressoDB will be easier for such user(s) to implement. In this scenario, we imagine that the other members of the collaboration are not able to set up an EspressoDB project themselves, but they could learn how to interface with it. The ease of using this framework AFTER the tables are defined is important in the context of coordinating collaborative projects. This scenario describes our own collaboration (except that one of the authors of this paper is also familiar enough with Django to have created EspressoDB) and it is the most common scenario in our field.  We assume our field is not unique and that many other collaborations that utilize supercomputers would find this database software useful for reducing the amount of human effort required to manage their complex calculations.

**Question:** Do you think that this process might be too difficult without prior knowledge of Django and if so, do you have suggestions on what to improve to make it more accessible?

> A potential issue I see with this framework is that it seems to bring together system administration and data processing in a way that I'm not sure is ideal. Perhaps some explanation of how LatteDB (or a hypothetical example instead) is implemented could be sufficient. I'm left wondering, for example, who manages the system and who uses it to do computational work. If everyone using the system (i.e. writing data processing code) needs to know Django and/or web development, the applicability of the framework may be somewhat limited. The paper and/or docs might benefit from some explanation of a reasonable workflow involving multiple team members with different roles (e.g. non-scientific administrator, computation-focused programmers, scientists pulling data to do local analyses, etc.).

The answer to this question is also addressed above.

We envision that a minimum of one collaborator has sufficient familiarity with Python and databases that they can implement the EspressoDB project.  We use ourselves as a good example.  One of us would have had no chance of efficiently creating EspressoDB or creating LatteDB (an example EspressoDB project).  However, with a little guidance from the main author of EspressoDB, he was able to modify existing Python scripts currently managing computations on an INCITE Allocation on Summit to begin using LatteDB.  These scripts are the ones we reference in the paper located in the `scripts` folder, with `lattedb` as part of the name of the script in the git repo located at

[`https://github.com/callat-qcd/nucleon_elastic_FF`](https://github.com/callat-qcd/nucleon_elastic_FF)

These scripts populate tables for which a GUI interface has been created, easily allowing users to view progress on creating and saving of data files on the file system and the tape backup system.  We provide a screenshots of this GUI below (it is not public as the results being managed are still in progress and have not been published yet).

https://github.com/openjournals/joss-reviews/issues/2007
