### 2020-02-13 callat-qcd/espressodb#50 @ckoerber
[JOSS Review] Reply to referee @remram44
Hello @remram44,

We would like to thank you for your detailed reply. We have split our response into two parts (according to the divider in your original post).

> there is a branch v1.0.0 with the same name as the tag v1.0.0...

Good point. The branch `v1.0.0` has been removed and we will stick to the following convention: Branches will consistently start with a `v` (like `v1.0.0`) and as soon as they are merged into master, a tag with the same name will be created so we can delete the merged branch.

> the blackmagicsorcery module stands out a bit, especially now that it just re-exports Python's standard re module (since 31c69150). You should probably just import re directly.

You are the first person to find this Easter egg :sweat_smile:. We believe this is not too critical as all import statements are either
```Python
from espressodb.base.utilities import blackmagicsorcery as re
```
or
```Python
from espressodb.base.utilities.blackmagicsorcery import search
```

> Regarding the paper markup...

Indeed: Django monospace will be fixed and since "Leadership computing facilities" is a name (see, e.g., https://www.olcf.ornl.gov) we will capitalize it instead.

https://github.com/openjournals/joss-reviews/issues/2007
