### 2020-02-14 466ddc5
andre's suggested change to the final LatteDB section
