+++ paper.md
@@ -84,22 +84,26 @@ The website generated by LatteDB used by CalLat can be found at [https://ithems.
 
 Summit at OLCF is disruptively fast compared to previous generations of leadership-class computers.  There are two challenges which are both critical to address for near-exascale computers such as Summit, which will become more important in the exascale era:
 
-1. _Dependent task generation and data processing_
-2. _Efficient bundling and management of intdependent tasks_
+1. _Efficient bundling and management of intdependent tasks_.
+2. _Dependent task generation and data processing_;
 
-The computational model for lattice QCD computations can be organized as a directed multi-graph:
+Using lattice QCD as a specific example, the computations can be organized as a directed multi-graph:
 a single calculation requires tens-of-thousands to millions of independent MPI tasks to be completed. These tasks, while running independently, have nested and chained interdependencies (the output of some tasks are part of the input for other tasks).
 Several such complete computations must be performed to extract final answers.
-However, LCFs prohibit the submission of millions of small tasks to their supercomputers (clogged queues, overtaxed service nodes, etc.).
-It is imperative to have a task manager capable of bundling many tasks into large jobs while distributing the work to various components of the heterogeneous nodes;
-To keep the nodes from going idle, the jobs must be backfilled while running with the next set of available tasks.  Keeping track of the tasks, optimizing the order of tasks and ensuring no work is repeated requires a task manager that understands all these dependencies and the uniqueness of each task.  Software to track and manage such a computational model at runtime, which does not require in-depth knowledge of, e.g., managing databases, does not currently exist.  
-
 As a specific example, CalLat creates petabytes of temporary files that are written to the scratch file system, used for subsequent computations and ultimately processed down to hundreds of terabytes that are saved for analysis.
 It is essential to track the status of these files in real-time to identify corrupt, missing, or purgeable files.
 
+Understandably, LCFs prohibit the submission of millions of small tasks to their supercomputers (clogged queues, overtaxed service nodes, etc.).
+It is therefore imperative to have a task manager capable of bundling many tasks into large jobs while distributing the work to various components of the heterogeneous nodes;
+To keep the nodes from going idle, the jobs must be backfilled while running with the next set of available tasks (item 1).
 Members of CalLat are addressing the task bundling through the creation of job management software, [METAQ](https://github.com/evanberkowitz/metaq) [@Berkowitz:2017vcp], and ``MPI_JM`` [@Berkowitz:2018gqe; @Berkowitz:2017xna], while
 LatteDB is designed to address the dependent task generation.
 A future feature of LatteDB  is integration with ``MPI_JM``.
+
+For the second item, keeping track of the tasks, optimizing the order of tasks and ensuring no work is repeated requires a task manager that understands all these dependencies and the uniqueness of each task.  Software to track and manage such a computational model at runtime, which does not require in-depth knowledge of, e.g., managing databases, does not currently exist, which motivated the creation of EspressoDB and LatteDB.
+
+
+
 
 
 # Acknowledgements
