+++ paper.md
@@ -84,10 +84,10 @@ The website generated by LatteDB used by CalLat can be found at [https://ithems.
 
 Summit at OLCF is disruptively fast compared to previous generations of leadership-class computers.  There are two challenges which are both critical to address for near-exascale computers such as Summit, which will become more important in the exascale era:
 
-1. _Efficient bundling and management of intdependent tasks_.
+1. _Efficient bundling and management of independent tasks_.
 2. _Dependent task generation and data processing_;
 
-Using lattice QCD as a specific example, the computations can be organized as a directed multi-graph:
+Using lattice QCD as a specific example, the computations can be organized as a directed multigraph:
 a single calculation requires tens-of-thousands to millions of independent MPI tasks to be completed. These tasks, while running independently, have nested and chained interdependencies (the output of some tasks are part of the input for other tasks).
 Several such complete computations must be performed to extract final answers.
 As a specific example, CalLat creates petabytes of temporary files that are written to the scratch file system, used for subsequent computations and ultimately processed down to hundreds of terabytes that are saved for analysis.
