### 2020-02-14 #2007 @ckoerber
Hello @gkthiruvathukal,  @ixjlyons, and @remram44,

We would like to thank you for your feedback and patience. We have posted detailed responses to both referee replies as issues on the EspressoDB repo

* [[JOSS Review] Reply to referee @ixjlyons](https://github.com/callat-qcd/espressodb/issues/49)
* [[JOSS Review] Reply to referee @remram44](https://github.com/callat-qcd/espressodb/issues/50)

Furthermore, we have created a [Pull Request which contains the updates](https://github.com/callat-qcd/espressodb/pull/48) to the paper, documentation, and new features we have introduced to address concerns made by the referees.

We intend to merge this branch into master once the second review iteration is finalized. Is this in accordance with the JOSS guidelines?

Feel free to contact us if you have any questions.

Best regards,

@cchang5, @ckoerber, @walkloud
