### 2020-02-17 callat-qcd/espressodb#49 @ixjlyons
> Question: Do you think that this process might be too difficult without prior knowledge of Django and if so, do you have suggestions on what to improve to make it more accessible?

Given the comments here and the additions to the paper and docs, I think the usage scenarios are much more clear, so this is less of a concern. My main issue was that expectations should be apparent to anyone thinking about starting an EspressoDB project.

Thank you for the thorough response. Commenting over in the main review issue now...
