### 2020-02-17 #2007 @ixjlyons
The authors have addressed my comments thoroughly. I looked over the updated and rendered paper from the v1.1.0 branch and updated my checklist. I recommend the submission be accepted.
