### 2020-02-18 #2007 @remram44
I completed my checklist after checking out https://github.com/callat-qcd/espressodb/pull/48, thanks!

Note that the list of workflow systems I knew off the top of my head might not be the most relevant systems (as I said, I used to work on one of them, but I've been out of this space for a few years. In particular, I don't think those tools are tailored to HPC. REANA and Galaxy are other names I remember).

Also there are probably academic publications for some systems which might be better citations than the URLs (Taverna has [this page](https://taverna.incubator.apache.org/community/cite) for example).
