### 2020-02-19 16c74d8
Added model reference in the quick docs
