+++ doc-src/example/project-creation/views.md
@@ -2,12 +2,12 @@
 
 Views are Python objects which provide information for html templates.
 E.g., they query the database and return strings to be rendered in your browser.
-When you enter an url in your browser, django figures out which Python object to call which knows which template to render.
+When you enter an url in your browser, Django figures out which Python object to call which knows which template to render.
 
 EspressoDB implements default views like the project homepage called which is rendered by the `index.html` template.
 
 ## Adjusting the index template
-When looking for the index template, django first checks your project template directories, e.g., `my_project/hamiltonian/templates/` and then enters EspressoDB's template dirs.
+When looking for the index template, Django first checks your project template directories, e.g., `my_project/hamiltonian/templates/` and then enters EspressoDB's template dirs.
 It renders the first template which matches the specification.
 Thus, when you create `my_project/hamiltonian/templates/index.html`, you will overwrite the default index page.
 
@@ -22,7 +22,7 @@ Here is an example of how to overwrite the index page.
     <p>
         See also the
         <a href="https://docs.djangoproject.com/en/dev/ref/templates/language/">
-            the django template documentation
+            the Django template documentation
         </a>
         how templates work.
     </p>
@@ -43,7 +43,7 @@ Most importantly, for this template are the texts `{% extends 'base.html' %}`, w
 Thus the link navbar and further html content will be present without you having to write anything.
 The `{% block content %}` and `{% endblock %}` denotes that within the `base.html` template, the content between the same exact blocks will be replaced with what you want to render.
 
-See also [the django docs for more information on templates](https://docs.djangoproject.com/en/dev/ref/templates/language/).
+See also [the Django docs for more information on templates](https://docs.djangoproject.com/en/dev/ref/templates/language/).
 
 Last but not least, EspressoDB comes with a few default css and javascript packages like [KaTeX](https://katex.org) to render equations and [Bootstrap 4](https://getbootstrap.com/docs/4.0/getting-started/introduction/) for having nice looking responsive webpages with minimal effort.
 
@@ -85,7 +85,7 @@ class HamiltonianStatusView(TemplateView):
 ```
 
 * The `TemplateView` class is a class with streamlines generating views.
-* The `template_name = "status.html"` tells django to look for a file called `status.html` to render this view
+* The `template_name = "status.html"` tells Django to look for a file called `status.html` to render this view
 * The `get_context_data` method provides additional information to the rendering. E.g., that the model we are using is a `ContactHamiltonian`.
     We overload the default method to guarantee that we do not eliminate other needed information.
 
@@ -108,7 +108,7 @@ To do so, create the file `my_project/hamiltonian/templates/status.html` and add
 The `{{model}}` now make use of the `context` parameter we have added (and will be rendered by the `str(ContactHamiltonian)` Python method).
 
 ### Update the urls
-To view this page online, we have to let django know where to find it.
+To view this page online, we have to let Django know where to find it.
 This is done by adjusting `my_project/hamiltonian/urls.py` to include the following lines
 ```python
 from django.urls import path
@@ -162,7 +162,7 @@ class HamiltonianStatusView(TemplateView):
 
         return df
 ```
-By default EspressoDB queries can be [converted to pandas DataFrames using django-pandas](https://github.com/chrisdev/django-pandas) this simplifies the logic of this code:
+By default EspressoDB queries can be [converted to Pandas `DataFrame`s using `django-pandas`](https://github.com/chrisdev/django-pandas) this simplifies the logic of this code:
 1. We get all Hamiltonians for the specified `self.model`
 2. We find all eigenvalues associated with the Hamiltonians
 3. For each hamiltonian (id), we count the number of associated entries
@@ -290,7 +290,7 @@ We have decided to not included it ourselves (meaning on you machine), as this a
[…]
