### 2020-02-19 1413d9d
Added non-break html char to Andre's name
