+++ paper.bib
@@ -234,3 +234,20 @@
                         JLAB-THY-10-1128",
       SLACcitation   = "%%CITATION = ARXIV:1001.3620;%%"
 }
+
+
+@article{10.1093/nar/gkt328,
+    author = {Wolstencroft, Katherine and Haines, Robert and Fellows, Donal and Williams, Alan and Withers, David and Owen, Stuart and Soiland-Reyes, Stian and Dunlop, Ian and Nenadic, Aleksandra and Fisher, Paul and Bhagat, Jiten and Belhajjame, Khalid and Bacall, Finn and Hardisty, Alex and Nieva de la Hidalga, Abraham and Balcazar Vargas, Maria P. and Sufi, Shoaib and Goble, Carole},
+    title = "{The Taverna workflow suite: designing and executing workflows of Web Services on the desktop, web or in the cloud}",
+    journal = {Nucleic Acids Research},
+    volume = {41},
+    number = {W1},
+    pages = {W557-W561},
+    year = {2013},
+    month = {05},
+    abstract = "{The Taverna workflow tool suite (http://www.taverna.org.uk) is designed to combine distributed Web Services and/or local tools into complex analysis pipelines. These pipelines can be executed on local desktop machines or through larger infrastructure (such as supercomputers, Grids or cloud environments), using the Taverna Server. In bioinformatics, Taverna workflows are typically used in the areas of high-throughput omics analyses (for example, proteomics or transcriptomics), or for evidence gathering methods involving text mining or data mining. Through Taverna, scientists have access to several thousand different tools and resources that are freely available from a large range of life science institutions. Once constructed, the workflows are reusable, executable bioinformatics protocols that can be shared, reused and repurposed. A repository of public workflows is available at http://www.myexperiment.org. This article provides an update to the Taverna tool suite, highlighting new features and developments in the workbench and the Taverna Server.}",
+    issn = {0305-1048},
+    doi = {10.1093/nar/gkt328},
+    url = {https://doi.org/10.1093/nar/gkt328},
+    eprint = {https://academic.oup.com/nar/article-pdf/41/W1/W557/3822473/gkt328.pdf},
+}
