### 2020-02-19 callat-qcd/espressodb#41 @ckoerber
The no binary install now works with v1.1.0.

Thank you for your feedback @ixjlyons.
