### 2020-02-20 a505587
Updated __str__ of Base
* Only columns which are None are not displayed
* Made sure that __str__ of specialization is called. Thus one can 
overwrite __str__
