+++ espressodb/base/models.py
@@ -187,22 +187,27 @@ class Base(models.Model):
     def __str__(self) -> str:
         """Verbose description of instance name, parent and column values.
         """
-        specialization = self.specialization
-        kwargs = {
-            field.name: getattr(specialization, field.name)
-            for field in specialization.get_open_fields()
-            if not isinstance(field, models.ForeignKey)
-            and not isinstance(field, models.ManyToManyField)
-            and getattr(specialization, field.name)
-        }
-        base = (
-            f"[{specialization.__class__.mro()[1].__name__}]"
-            if type(specialization) != Base  # pylint: disable=C0123
-            else ""
-        )
-        info = ", ".join([f"{key}={val}" for key, val in kwargs.items() if val])
-        info_str = f"({info})" if info else ""
-        return f"{specialization.__class__.__name__}{base}{info_str}"
+        if self == self.specialization:
+            kwargs = {
+                field.name: getattr(self, field.name)
+                for field in self.get_open_fields()
+                if not isinstance(field, models.ForeignKey)
+                and not isinstance(field, models.ManyToManyField)
+                and getattr(self, field.name)
+            }
+            base = (
+                f"[{self.__class__.mro()[1].__name__}]"
+                if type(self) != Base  # pylint: disable=C0123
+                else ""
+            )
+            info = ", ".join(
+                [f"{key}={val}" for key, val in kwargs.items() if val is not None]
+            )
+            info_str = f"({info})" if info else ""
+            out = f"{self.__class__.__name__}{base}{info_str}"
+        else:
+            out = self.specialization.__str__()
+        return out
 
     def save(  # pylint: disable=W0221
         self, *args, save_instance_only: bool = False, **kwargs,
