### 2020-02-20 #2007 @ckoerber
Hello @gkthiruvathukal, @ixjlyons, and @remram44,

We are happy that EspressoDB was accepted for publication and would like to thank you for your time and efforts.

The tagged release is [v1.1.0](https://github.com/callat-qcd/espressodb/releases/tag/v1.1.0) and the zenodo doi is 10.5281/zenodo.3676074 for all versions and 10.5281/zenodo.3677432 for the one associated with the tag. We have verified that the Zenodo metadata matches the JOSS metadata and added ORCIDs.
