### 2020-02-21 #2007 @kyleniemeyer
@ckoerber actually, it looks like one of your citations to an arXiv preprint (Berkowitz 2018) has since been published, can you update the info? https://doi.org/10.1109/sc.2018.00058
