+++ paper.bib
@@ -38,26 +38,18 @@
  address = {Piscataway, NJ, USA},
 }
 
-@Misc{Berkowitz:2018gqe,
-  author        = "Berkowitz, Evan
-      and Clark, M.A.
-      and Gambhir, Arjun
-      and McElvain, Ken
-      and Nicholson, Amy
-      and Rinaldi, Enrico
-      and Vranas, Pavlos
-      and Walker-Loud, Andr{\'e}
-      and Chang, Chia Cheng
-      and Jo{\'o}, B{\'a}lint
-      and Kurth, Thorsten
-      and Orginos, Kostas",
-  title         = "{Simulating the weak death of the neutron in a femtoscale universe with near-Exascale computing}",
-  year          = "2018",
-  archiveprefix = "arXiv",
-  eprint        = "1810.01609",
-  primaryclass  = "hep-lat",
-  slaccitation  = "%%CITATION = ARXIV:1810.01609;%%",
-}
+@INPROCEEDINGS{8665785,
+author={E. {Berkowitz} and M. A. {Clark} and A. {Gambhir} and K. {McElvain} and A. {Nicholson} and E. {Rinaldi} and P. {Vranas} and A. {Walker-Loud} and C. C. {Chang} and B. {Joó} and T. {Kurth} and K. {Orginos}},
+booktitle={SC18: International Conference for High Performance Computing, Networking, Storage and Analysis},
+title={Simulating the Weak Death of the Neutron in a Femtoscale Universe with Near-Exascale Computing},
+year={2018},
+volume={},
+number={},
+pages={697-705},
+keywords={graphics processing units;lattice field theory;multiprocessing systems;neutrons;parallel machines;physics computing;quantum chromodynamics;Quantum Chromodynamics;QCD;protons;intrinsic properties;atomic nuclei;neutron lifetime;numerical solutions;Lattice Gauge Theory;time-to-solution;run-time autotuning;femtoscale universe;near-exascale computing;fundamental particle theory;nuclear observables;CORAL supercomputers;GPU resources;Neutrons;Lattices;Kernel;Physics;Protons;Graphics processing units;Mathematical model},
+doi={10.1109/SC.2018.00058},
+ISSN={null},
+month={Nov},}
 
 @misc{incite:2019,
     author = "Walker-Loud, Andr{\'e}
