### 2020-02-21 #2007 @ckoerber
Hello @kyleniemeyer,

Yes the second reference is now updated as well. The updates to the references (only changes) are on the master branch which is now ahead of the previous `v1.1.0` tag. Do we have to move or create a new tag for the submission (if so, do we have to updated the Zenodo doi as well)?
