### 2020-02-21 #2007 @kyleniemeyer
@ckoerber no, we aren't concerned with a minor update to the article—we archive that. I'll proceed with the existing Zenodo archive and version.
