### 2020-02-21 176f283
Update README.md
Added JOSS badge to repo
