### 2020-02-21 #2007 @kyleniemeyer
Congrats  @ckoerber & @cchang5 on your article's publication in JOSS!
